vscode tasks, fixing shared lib workaround, refactoring
