+++ .vscode/launch.json
@@ -4,22 +4,39 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
-        
+
+        // {
+        //     "request": "launch",
+        //     "name": "Debug Server Tests",
+        //     "type": "node",
+        //     "cwd": "${workspaceFolder}/server",
+        //     "env": {},
+        //     "runtimeExecutable": "/opt/homebrew/bin/deno",
+        //     "runtimeArgs": [
+        //         "test",
+        //         "--inspect-wait=0.0.0.0:9229",
+        //         "--allow-all",
+        //         "--env-file=./.env.local"
+        //     ],
+        //     "attachSimplePort": 9229
+        // },
+           
         {
             "request": "launch",
-            "name": "Launch Program",
+            "name": "Debug Run Server Tests",
             "type": "node",
-            "program": "${workspaceFolder}/main.ts",
-            "cwd": "${workspaceFolder}",
+            "cwd": "${workspaceFolder}/server",
             "env": {},
             "runtimeExecutable": "/opt/homebrew/bin/deno",
             "runtimeArgs": [
-                "run",
-                "--unstable",
-                "--inspect-wait",
-                "--allow-all"
+                "test",
+                "--allow-all",
+                "--coverage",
+                "--env-file=./.env.local"
             ],
-            "attachSimplePort": 9229
+            "console": "integratedTerminal",
+            "attachSimplePort": 9229,
+            "postDebugTask": "Generate Coverage Report"
         }
     ]
 }
