+++ server/supabase/functions/_shared/db.ts
@@ -1,7 +1,14 @@
+import "jsr:@std/dotenv/load";
+
 import { createClient } from 'jsr:@supabase/supabase-js@2'
 
 function getSupabaseClient() {
-    return createClient(Deno.env.get('SUPABASE_URL') ?? '', Deno.env.get('SUPABASE_SERVICE_ROLE_KEY') ?? '')
+    const supabaseUrl = Deno.env.get('SUPABASE_URL') ?? '';
+    const supabaseServiceRoleKey = Deno.env.get('SUPABASE_SERVICE_ROLE_KEY') ?? '';
+    if (!supabaseUrl || !supabaseServiceRoleKey) {
+        throw new Error('SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY must be set');
+    }
+    return createClient(supabaseUrl, supabaseServiceRoleKey)
 }
 
 export const supa = getSupabaseClient();
