+++ server/supabase/functions/users/deno.json
@@ -2,7 +2,6 @@
   "imports": {
     "@hono/hono": "jsr:@hono/hono@^4.7.2",
     "@supabase/supabase-js": "jsr:@supabase/supabase-js@^2.49.1",
-    "zod": "npm:zod@^3.24.2",
     "@shared/": "../../../../shared/"
   }
 }
