+++ shared/deno.json
@@ -1,17 +1,13 @@
 {
-  "name": "@scope/shared",
+  "name": "@armin/forum-shared",
+  "version": "0.1.0",
+  "license": "MIT",
+  "exports": "./mod.ts",
   "tasks": {
     "dev": "deno run --watch main.ts"
   },
   "imports": {
     "@std/assert": "jsr:@std/assert@1",
     "zod": "npm:zod@^3.24.2"
-
-  },
-  "version": "0.1.0",
-  "exports": {
-    "./schema/users": "./schema/users.ts",
-    "./schema/users.ts": "./schema/users.ts"
   }
-
 }
