+++ shared/deps.ts
@@ -1,0 +1,2 @@
+export * as z from "npm:zod"
+
